separated Todos component
