+++ app/add-todo.js
@@ -2,15 +2,19 @@
 
 import { useState } from 'react';
 
-export default function AddTodo({ addTodo, todosLength }) {
+export default function AddTodo({ todos, setTodos }) {
   const [todoText, setTodoText] = useState('');
 
+  const addTodo = (todo) => {
+    setTodos([...todos, todo])
+  }
+  
   const handleChange = (e) => {
     setTodoText(e.target.value);
   }
   const handleClick = () => {
     if (todoText.trim() != "") {
-        addTodo({"id": todosLength, "title": todoText.trim(), "done" : false})
+        addTodo({"id": todos.length, "title": todoText.trim(), "done" : false})
         setTodoText('');
     }
   }
